KANI alias & dependency hell
- Aliases are now implemented, shows a list o
- I really need to make a full restructure of the rails feature because RailTraverse is handling a lot of the bloat right now and it's really messy, mostly because it depends on a JSON right now; maybe update database every three hours to make it global?
